Test schema changes
In order to test the pipeline, this change creates a schema change on the database and adds application code that depends on this change. Hitting the /healthcheck route will depend on the new table created by the new migration, update the record, and print the last healthcheck time. This should allow for interacting with the test/prod APIs as it moves through the pipeline to see that the change to both app/db have successfully deployed.