Add rails migration entry
This adds an entry point for doing rails migrations. This is similar to
the entry point for the rails service, except it does not need to start
rails or nginx, so made sense to separate these out.

Additional changes:
- Changed the start services script to be named rails_entry.sh and use
a wait for db line to match how we're doing other services
- Simplified required newman params. Since we require https, seemed
overkill to require this as a param